Crate rustc_error_messages

Re-exports

pub use fluent_generated as fluent;

Macros

langid

Structs

FluentArgs

A map of arguments passed from the code to the localization to be used for message formatting.

LanguageIdentifier

LanguageIdentifier is a core struct representing a Unicode Language Identifier.

MultiSpan

A collection of Spans.

SpanLabel

A span together with some additional data.

Enums

DiagnosticMessage

Abstraction over a message in a diagnostic to support both translatable and non-translatable diagnostic messages.

FluentError

Core error type for Fluent runtime system.

FluentValue

The FluentValue enum represents values which can be formatted to a String.

SubdiagnosticMessage

Abstraction over a message in a subdiagnostic (i.e. label, note, help, etc) to support both translatable and non-translatable diagnostic messages.

TranslationBundleError

Statics

DEFAULT_LOCALE_RESOURCES

Functions

fallback_fluent_bundle

Return the default FluentBundle with standard “en-US” diagnostic messages.

fluent_bundle

Returns Fluent bundle with the user’s locale resources from $sysroot/share/locale/$requested_locale/*.ftl.

new_bundle 🔒

Type Definitions

FluentBundle

FluentId 🔒

Identifier for the Fluent message/attribute corresponding to a diagnostic message.

LazyFallbackBundle

Type alias for the result of fallback_fluent_bundle - a reference-counted pointer to a lazily evaluated fluent bundle.