Enum rustc_error_messages::DiagnosticMessage

pub enum DiagnosticMessage {
    Str(String),
    FluentIdentifier(Cow<'static, str>, Option<Cow<'static, str>>),
}

Abstraction over a message in a diagnostic to support both translatable and non-translatable diagnostic messages.

Intended to be removed once diagnostics are entirely translatable.

Variants

Str(String)

Non-translatable diagnostic message.

FluentIdentifier(Cow<'static, str>, Option<Cow<'static, str>>)

Identifier for a Fluent message (with optional attribute) corresponding to the diagnostic message.

https://projectfluent.org/fluent/guide/hello.html https://projectfluent.org/fluent/guide/attributes.html

Implementations

impl DiagnosticMessage

pub fn with_subdiagnostic_message(&self, sub: SubdiagnosticMessage) -> Self

Given a SubdiagnosticMessage which may contain a Fluent attribute, create a new DiagnosticMessage that combines that attribute with the Fluent identifier of self.

If the SubdiagnosticMessage is non-translatable then return the message as a DiagnosticMessage.

If self is non-translatable then return self’s message.

pub fn expect_str(&self) -> &str

Returns the String contained within the DiagnosticMessage::Str variant, assuming that this diagnostic message is of the legacy, non-translatable variety. Panics if this assumption does not hold.

Don’t use this - it exists to support some places that do comparison with diagnostic strings.