Struct rustc_error_messages::MultiSpan

pub struct MultiSpan {
    pub(crate) primary_spans: Vec<Span>,
    pub(crate) span_labels: Vec<(Span, DiagnosticMessage)>,
}

A collection of Spans.

Spans have two orthogonal attributes:

They can be primary spans. In this case they are the locus of the error, and would be rendered with ^^^.

They can have a label. In this case, the label is written next to the mark in the snippet when we render.

Fields
primary_spans: Vec<Span>span_labels: Vec<(Span, DiagnosticMessage)>
Implementations

impl MultiSpan

pub fn new() -> MultiSpan

pub fn from_span(primary_span: Span) -> MultiSpan

pub fn from_spans(vec: Vec<Span>) -> MultiSpan

pub fn push_span_label( &mut self, span: Span, label: impl Into<DiagnosticMessage> )

pub fn primary_span(&self) -> Option<Span>

Selects the first primary span (if any).

pub fn primary_spans(&self) -> &[Span]

Returns all primary spans.

pub fn has_primary_spans(&self) -> bool

Returns true if any of the primary spans are displayable.

pub fn is_dummy(&self) -> bool

Returns true if this contains only a dummy primary span with any hygienic context.

pub fn replace(&mut self, before: Span, after: Span) -> bool

Replaces all occurrences of one Span with another. Used to move Spans in areas that don’t display well (like std macros). Returns whether replacements occurred.

pub fn pop_span_label(&mut self) -> Option<(Span, DiagnosticMessage)>

pub fn span_labels(&self) -> Vec<SpanLabel>

Returns the strings to highlight. We always ensure that there is an entry for each of the primary spans – for each primary span P, if there is at least one label with span P, we return those labels (marked as primary). But otherwise we return SpanLabel instances with empty labels.

pub fn has_span_labels(&self) -> bool

Returns true if any of the span labels is displayable.

pub fn clone_ignoring_labels(&self) -> Self

Clone this MultiSpan without keeping any of the span labels - sometimes a MultiSpan is to be re-used in another diagnostic, but includes span_labels which have translated messages. These translated messages would fail to translate without their diagnostic arguments which are unlikely to be cloned alongside the Span.